Refactored each api call into seperate files
